Move captcha code to its own file
